+++ docs_web/android/app/src/main/kotlin/com/example/docs_web/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.docs_web
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}